+++ src/main/java/com/braintreegateway/VenmoProfileData.java
@@ -3,11 +3,13 @@ package com.braintreegateway;
 import com.braintreegateway.util.NodeWrapper;
 
 public class VenmoProfileData {
-    private String username;
+    private Address billingAddress;
+    private String email;
     private String firstName;
     private String lastName;
     private String phoneNumber;
-    private String email;
+    private Address shippingAddress;
+    private String username;
 
     public VenmoProfileData(NodeWrapper node) {
         this.username = node.findString("username");
@@ -15,6 +17,16 @@ public class VenmoProfileData {
         this.lastName = node.findString("last-name");
         this.phoneNumber = node.findString("phone-number");
         this.email = node.findString("email");
+
+        NodeWrapper billingAddress = node.findFirst("billing-address");
+        if (billingAddress != null) {
+            this.billingAddress = new Address(billingAddress);
+        }
+
+        NodeWrapper shippingAddress = node.findFirst("shipping-address");
+        if (shippingAddress != null) {
+            this.shippingAddress = new Address(shippingAddress);
+        }
     }
 
     public String getUsername() {
@@ -35,5 +47,13 @@ public class VenmoProfileData {
 
     public String getEmail() {
         return email;
+    }
+
+    public Address getBillingAddress() {
+        return billingAddress;
+    }
+
+    public Address getShippingAddress() {
+        return shippingAddress;
     }
  }